refactor: moved SignUpController to their own production file
