+++ tests/presentation/controllers/signup-controller.spec.ts
@@ -1,5 +1,5 @@
-import { IController, IHttpRequest, IHttpResponse } from '@/../../src/presentation/protocols'
 import { badRequest } from '@/../../src/presentation/helpers'
+import { SignUpController } from '@/../../src/presentation/controller'
 import faker, { fake } from 'faker'
 
 
@@ -19,27 +19,6 @@ const makeSut = () => {
     }
 }
 
-export class SignUpController implements IController{
-    async handle(request: IHttpRequest): Promise<IHttpResponse>{
-        try {
-            const paramsRequired = ["username", "email", "password", "passwordConfirmation"]
-            for (const params of paramsRequired) {
-                if (!request.body[params]) {
-                    return badRequest(params)
-                }
-            }
-            return {
-                statusCode: 200,
-                body: ""
-            }
-        } catch (error) {
-            return {
-                statusCode: 500,
-                body: error
-            }
-        }
-    }
-}
